refactor: use decimal instead of float for better precision while storing amount

+++ internal/migrations/1763416987_create_accounts.sql
@@ -3,6 +3,6 @@
 
 CREATE TABLE IF NOT EXISTS accounts (
     id TEXT PRIMARY KEY,
-    balance REAL NOT NULL DEFAULT 0
+    balance NUMERIC(23, 5) NOT NULL DEFAULT 0
 );
